requirements.txt
tensorflow-cpu==2.12.0
streamlit==1.23.1
